Pull instruction definitions out of interpreter.cpp.
